pyproject.toml
[tool.poetry]
name = "teachservice"
version = "0.1.0"
description = ""
authors = ["Amir Mustafaev <amir.mustafaev777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
